Move kbpackets out
