+++ go/kbcrypto/util.go
@@ -1,0 +1,17 @@
+// Copyright 2018 Keybase, Inc. All rights reserved. Use of
+// this source code is governed by the included BSD license.
+
+package kbcrypto
+
+import (
+	"bytes"
+	"crypto/hmac"
+)
+
+func FastByteArrayEq(a, b []byte) bool {
+	return bytes.Equal(a, b)
+}
+
+func SecureByteArrayEq(a, b []byte) bool {
+	return hmac.Equal(a, b)
+}
