+++ go/libkb/kbpackets_test.go
@@ -6,6 +6,7 @@ package libkb
 import (
 	"testing"
 
+	"github.com/keybase/client/go/kbcrypto"
 	"github.com/stretchr/testify/require"
 )
 
@@ -61,30 +62,7 @@ NGQwYjE5NTMwMzIwOWNiZDgyYWZhN2ZjNmRlZDE4YjQ5YjdiNmIyIiwic2Vxbm8iOjQsInRhZyI6
 InNpZ25hdHVyZSJ9o3NpZ8RAbe4i3mDpfo1ECOcd0XU1JE7lRgkPUHQq7WLEEh5LbO5IAZHSm2zY
 tuX4LPcEa+72KyrsweuAJravU8SjgL/gAKhzaWdfdHlwZSCjdGFnzQICp3ZlcnNpb24B
 `)
-	require.IsType(t, err, FishyMsgpackError{}, "info=%+v, err+%+v", info, err)
-}
-
-type testPacketable struct{}
-
-func (*testPacketable) GetTagAndVersion() (PacketTag, PacketVersion) {
-	return TagSignature, KeybasePacketV1
-}
-
-// Guard against unexpected codec encoding changes, in particular for
-// ints.
-func TestHardcodedPacketEncode(t *testing.T) {
-	var nilPtr *testPacketable
-	p, err := newKeybasePacket(nilPtr)
-	require.NoError(t, err)
-
-	p.Hash = nil
-
-	bytes, err := p.encode()
-	require.NoError(t, err)
-	// In particular, {0xcd, 0x2, 0x2} shouldn't change to
-	// {0xd1, 0x2, 0x2}.
-	expectedBytes := []byte{0x83, 0xa4, 0x62, 0x6f, 0x64, 0x79, 0xc0, 0xa3, 0x74, 0x61, 0x67, 0xcd, 0x2, 0x2, 0xa7, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x1}
-	require.Equal(t, expectedBytes, bytes)
+	require.IsType(t, err, kbcrypto.FishyMsgpackError{}, "info=%+v, err+%+v", info, err)
 }
 
 // This is a regression test for
